urlextraction/requirements.txt
langchain
google-generativeai
faiss-cpu
unstructured
sentence-transformers
numpy
django
